Added Spicy grilled cheese sandwiches to the recipe list
Contains the procedures on how to cook spicy grilled cheese sandwhiches all on a new page that has been linked to the index page.

Spicy Grilled cheese are wonderful, and this is somethign I have never tried, but it is something that I would like to, and I think a lot of people would find it interesting to try out too.